Broke State up into several modules in Types
