+++ parsley-core/src/ghc/Parsley/Internal/Backend/Machine/LetRecBuilder.hs
@@ -13,7 +13,8 @@ import Language.Haskell.TH.Syntax                   (Q, unTypeQ, unsafeTExpCoerc
 import Language.Haskell.TH.Syntax                   (unTypeCode, unsafeCodeCoerce, Exp(VarE, LetE), Dec(FunD), Clause(Clause), Body(NormalB))
 #endif
 import Parsley.Internal.Backend.Machine.LetBindings (LetBinding(..), Binding, Regs)
-import Parsley.Internal.Backend.Machine.State       (QSubRoutine, qSubRoutine, Func)
+import Parsley.Internal.Backend.Machine.Types.State (QSubRoutine, qSubRoutine, Func)
+
 import Parsley.Internal.Common.Utils                (Code)
 
 import Data.Dependent.Map as DMap (DMap, (!), map, toList, traverseWithKey)
